+++ cli/commands/toolchain/lintoko.ts
@@ -10,12 +10,12 @@ let cacheDir = path.join(globalCacheDir, "lintoko");
 
 export let repo = "caffeinelabs/lintoko";
 
-export let getLatestReleaseTag = async () => {
-  return toolchainUtils.getLatestReleaseTag(repo);
+export let getLatestReleaseTag = async ({ prerelease = false } = {}) => {
+  return toolchainUtils.getLatestReleaseTag(repo, { prerelease });
 };
 
-export let getReleases = async () => {
-  return toolchainUtils.getReleases(repo);
+export let getReleases = async ({ prerelease = false } = {}) => {
+  return toolchainUtils.getReleases(repo, { prerelease });
 };
 
 export let isCached = (version: string) => {